requirements.txt
streamlit==1.44.1
pandas==2.2.3
plotly==5.24.1
groq==0.18.0
python-dotenv==1.0.1
